Move model extension functions into data layer
