+++ feature/books/src/main/java/com/shahin/feature/books/data/extensions/BookItemExtensions.kt
@@ -1,4 +1,4 @@
-package com.shahin.feature.books.extensions
+package com.shahin.feature.books.data.extensions
 
 import com.shahin.core.database.books.model.BookEntity
 import com.shahin.core.network.books.model.BookItem
